Struts lab

+++ src/main/java/org/superbiz/struts/UserService.java
@@ -17,13 +17,15 @@
 */
 package org.superbiz.struts;
 
+import org.springframework.stereotype.Component;
+
 import java.util.List;
 
 public interface UserService {
 
     public void add(User user);
 
-    public User find(int id);
+    public User find(long id);
 
     public List<User> findAll();
 }
